+++ renku/core/workflow/plan.py
@@ -20,7 +20,9 @@
 import itertools
 from datetime import datetime, timedelta
 from pathlib import Path
-from typing import Dict, Generator, List, Optional, Set, Tuple, Union, cast, overload
+from typing import Any, Dict, Generator, List, Optional, Set, Tuple, Union, cast, overload
+
+from pydantic import validate_arguments
 
 from renku.command.command_builder import inject
 from renku.command.format.workflow import WORKFLOW_FORMATS
@@ -32,7 +34,7 @@ from renku.core.interface.project_gateway import IProjectGateway
 from renku.core.util import communication
 from renku.core.util.datetime8601 import local_now
 from renku.core.util.git import get_git_user
-from renku.core.util.os import are_paths_related, get_relative_paths, safe_read_yaml
+from renku.core.util.os import are_paths_related, get_relative_paths
 from renku.core.util.util import NO_VALUE, NoValueType
 from renku.core.workflow.concrete_execution_graph import ExecutionGraph
 from renku.core.workflow.value_resolution import CompositePlanValueResolver, ValueResolver
@@ -89,6 +91,7 @@ def get_derivative_chain(
 
 
 @inject.autoparams()
+@validate_arguments(config=dict(arbitrary_types_allowed=True))
 def search_workflows(name: str, plan_gateway: IPlanGateway) -> List[str]:
     """Get all the workflows whose Plan.name start with the given name.
 
@@ -103,7 +106,8 @@ def search_workflows(name: str, plan_gateway: IPlanGateway) -> List[str]:
 
 
 @inject.autoparams()
-def list_workflows(plan_gateway: IPlanGateway, format: str, columns: List[str]):
+@validate_arguments(config=dict(arbitrary_types_allowed=True))
+def list_workflows(plan_gateway: IPlanGateway, format: str, columns: str):
     """List or manage workflows with subcommands.
 
     Args:
@@ -128,6 +132,7 @@ def list_workflows(plan_gateway: IPlanGateway, format: str, columns: List[str]):
 
 
 @inject.autoparams("plan_gateway", "activity_gateway")
+@validate_arguments(config=dict(arbitrary_types_allowed=True))
 def show_workflow(
     name_or_id: str, plan_gateway: IPlanGateway, activity_gateway: IActivityGateway, with_metadata: bool = False
 ):
@@ -185,6 +190,7 @@ def show_workflow(
 
 
 @inject.autoparams("plan_gateway")
+@validate_arguments(config=dict(arbitrary_types_allowed=True))
 def remove_plan(name_or_id: str, force: bool, plan_gateway: IPlanGateway, when: datetime = local_now()):
     """Remove the workflow by its name or id.
 
@@ -230,6 +236,7 @@ def remove_plan(name_or_id: str, force: bool, plan_gateway: IPlanGateway, when: 
 
 
 @inject.autoparams()
+@validate_arguments(config=dict(arbitrary_types_allowed=True))
 def edit_workflow(
     name: str,
     new_name: Optional[str],
@@ -338,21 +345,22 @@ def edit_workflow(
 
 
 @inject.autoparams()
+@validate_arguments(config=dict(arbitrary_types_allowed=True))
 def compose_workflow(
     name: str,
-    description: str,
-    mappings: List[str],
-    defaults: List[str],
-    links: List[str],
-    param_descriptions: List[str],
+    description: Optional[str],
+    mappings: Optional[List[str]],
+    defaults: Optional[List[str]],
+    links: Optional[List[str]],
+    param_descriptions: Optional[List[str]],
     map_inputs: bool,
     map_outputs: bool,
     map_params: bool,
     link_all: bool,
-    keywords: List[str],
-    steps: List[str],
-    sources: List[str],
-    sinks: List[str],
+    keywords: Optional[List[str]],
+    steps: Optional[List[str]],
+    sources: Optional[List[str]],
+    sinks: Optional[List[str]],
     creators: Optional[List[Person]],
     activity_gateway: IActivityGateway,
     plan_gateway: IPlanGateway,
@@ -362,19 +370,19 @@ def compose_workflow(
 
     Args:
         name(str): Name of the new composed Plan.
-        description(str): Description for the Plan.
-        mappings(List[str]): Mappings between parameters of this and child Plans.
-        defaults(List[str]): Default values for parameters.
-        links(List[str]): Links between parameters of child Plans.
-        param_descriptions(List[str]): Descriptions of parameters.
+        description(Optional[str]): Description for the Plan.
+        mappings(Optional[List[str]]): Mappings between parameters of this and child Plans.
+        defaults(Optional[List[str]]): Default values for parameters.
+        links(Optional[List[str]]): Links between parameters of child Plans.
+        param_descriptions(Optional[List[str]]): Descriptions of parameters.
         map_inputs(bool): Whether or not to automatically expose child inputs.
         map_outputs(bool): Whether or not to automatically expose child outputs.
         map_params(bool): Whether or not to automatically expose child parameters.
         link_all(bool): Whether or not to automatically link child steps' parameters.
-        keywords(List[str]): Keywords for the Plan.
-        steps(List[str]): Child steps to include.
-        sources(List[str]): Starting files when automatically detecting child Plans.
-        sinks(List[str]): Ending files when automatically detecting child Plans.
+        keywords(Optional[List[str]]): Keywords for the Plan.
+        steps(Optional[List[str]]): Child steps to include.
+        sources(Optional[List[str]]): Starting files when automatically detecting child Plans.
+        sinks(Optional[List[str]]): Ending files when automatically detecting child Plans.
         creators(Optional[List[Person]]): Creator(s) of the composite plan.
         activity_gateway(IActivityGateway): Injected activity gateway.
         plan_gateway(IPlanGateway): Injected plan gateway.
@@ -504,12 +512,16 @@ def compose_workflow(
 
 
 @inject.autoparams()
+@validate_arguments(config=dict(arbitrary_types_allowed=True))
 def export_workflow(
     name_or_id,
     plan_gateway: IPlanGateway,
     format: str,
     output: Optional[Union[str, Path]],
-    values: Optional[str],
+    values: Optional[Dict[str, Any]],
+    basedir: Optional[str],
+    resolve_paths: Optional[bool],
+    nest_workflows: Optional[bool],
 ):
     """Export a workflow to a given format.
 
@@ -518,10 +530,23 @@ def export_workflow(
         plan_gateway(IPlanGateway): The injected Plan gateway.
         format(str): Format to export to.
         output(Optional[str]): Output path to store result at.
-        values(Optional[str]): Path to values file to apply before export.
+        values(Optional[Dict[str,Any]]): Parameter names and values to apply before export.
+        basedir(Optional[str]): The base path prepended to all paths in the exported workflow,
+            if None it defaults to the absolute path of the renku project.
+        resolve_paths(Optional[bool]): Resolve all symlinks and make paths absolute, defaults to True.
+        nest_workflows(Optional[bool]): Whether to try to nest all workflows into one specification and file or not,
+            defaults to False.
     Returns:
         The exported workflow as string.
     """
+
+    if resolve_paths is None:
+        resolve_paths = True
+
+    if basedir is None:
+        basedir_path = project_context.path
+    elif isinstance(basedir, str):
+        basedir_path = Path(basedir)
 
     workflow = plan_gateway.get_by_name_or_id(name_or_id)
 
@@ -535,8 +560,7 @@ def export_workflow(
         output_path = cast(Path, output)
 
     if values:
-        parsed_values = safe_read_yaml(values)
-        rv = ValueResolver.get(workflow, parsed_values)
+        rv = ValueResolver.get(workflow, values)
         workflow = rv.apply()
         if rv.missing_parameters:
             communication.warn(
@@ -547,7 +571,14 @@ def export_workflow(
     from renku.core.plugin.workflow import workflow_converter
 
     converter = workflow_converter(format)
-    return converter(workflow=workflow, basedir=project_context.path, output=output_path, output_format=format)
+    return converter(
+        workflow=workflow,
+        basedir=basedir_path,
+        output=output_path,
+        output_format=format,
+        resolve_paths=resolve_paths,
+        nest_workflows=nest_workflows,
+    )
 
 
 def _lookup_paths_in_paths(lookup_paths: List[str], target_paths: List[str]):
@@ -587,6 +618,7 @@ def _lookup_paths_in_paths(lookup_paths: List[str], target_paths: List[str]):
 
 
 @inject.autoparams()
+@validate_arguments(config=dict(arbitrary_types_allowed=True))
 def visualize_graph(
     sources: List[str],
     targets: List[str],
@@ -628,7 +660,8 @@ def visualize_graph(
 
 
 @inject.autoparams()
-def workflow_inputs(activity_gateway: IActivityGateway, paths: List[str] = None):
+@validate_arguments(config=dict(arbitrary_types_allowed=True))
+def workflow_inputs(activity_gateway: IActivityGateway, paths: Optional[List[str]] = None):
     """Get inputs used by workflows.
 
     Args:
@@ -647,7 +680,8 @@ def workflow_inputs(activity_gateway: IActivityGateway, paths: List[str] = None)
 
 
 @inject.autoparams()
-def workflow_outputs(activity_gateway: IActivityGateway, paths: List[str] = None):
+@validate_arguments(config=dict(arbitrary_types_allowed=True))
+def workflow_outputs(activity_gateway: IActivityGateway, paths: Optional[List[str]] = None):
     """Get inputs used by workflows.
 
     Args: