+++ renku/ui/service/serializers/workflows.py
@@ -16,10 +16,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Renku service workflow serializers."""
+from enum import Enum
+
 from marshmallow import Schema, fields
 from marshmallow_oneofschema import OneOfSchema
 
 from renku.domain_model.dataset import DatasetCreatorsJson
+from renku.ui.cli.utils.plugins import supported_formats
 from renku.ui.service.serializers.common import LocalRepositorySchema, RemoteRepositorySchema
 from renku.ui.service.serializers.rpc import JsonRPCResponse
 
@@ -191,3 +194,24 @@ class WorkflowPlansShowResponseRPC(JsonRPCResponse):
     """Response schema for plan show view."""
 
     result = fields.Nested(PlanSuperSchema)
+
+
+WorkflowExportFormatEnum = Enum(  # type: ignore
+    "WorkflowExportFormatEnum",
+    zip(supported_formats(), supported_formats()),
+)
+
+
+class WorkflowPlansExportRequest(LocalRepositorySchema, RemoteRepositorySchema):
+    """Request schema for exporting a plan."""
+
+    plan_id = fields.String(required=True)
+    # NOTE: the enum values (not names) are used to deserialize format, the result from deserialize is Enum
+    format = fields.Enum(WorkflowExportFormatEnum, missing=getattr(WorkflowExportFormatEnum, "cwl"), by_value=True)
+    values = fields.Dict(keys=fields.String(), missing=None)
+
+
+class WorkflowPlansExportResponseRPC(JsonRPCResponse):
+    """Response schema for exporting a plan."""
+
+    result = fields.String()